drupal/leaflet 10.5.2, drupal/imagemagick 3.6.0, drupal/image_effects: 3.5.0, drupal/file_mdm 3.0.0

+++ web/modules/contrib/image_effects/src/Component/ImageUtility.php
@@ -31,7 +31,7 @@ abstract class ImageUtility {
       if ($current_length === NULL) {
         return NULL;
       }
-      return (int) (str_replace('%', '', $length_specification) * 0.01 * $current_length);
+      return (int) (((float) str_replace('%', '', $length_specification)) * 0.01 * $current_length);
     }
     return (int) $length_specification;
   }
